backend/requirements.txt
fastapi==0.117.1
uvicorn==0.30.6
web3==6.20.1
pymongo==4.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic[email]==2.5.0
email-validator==2.1.0
